requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
python-dotenv
groq
passlib[bcrypt]
python-jose[cryptography]
pydantic[email]
